리뷰 반영 - Blackjack 은 컨트롤러이므로 view 객체에 의존하도록 변경
