--- src/main/kotlin/blackjack/Fetcher.kt
@@ -1,5 +1,0 @@
-package blackjack
-
-fun interface Fetcher<A, B> {
-    fun fetch(args: A): B
-}
