--- src/main/kotlin/blackjack/Printer.kt
@@ -1,5 +1,0 @@
-package blackjack
-
-fun interface Printer<A> {
-    fun print(args: A)
-}
